Fix runtime error when running with WEKA-3.8.5
The problem is that java11 has removed JAXB module which is used by Auto-WEKA.

I find that WEKA-3.8.5 has already included the necessary packages (jakarta.bine-api:2.3.3 and jaxb-runtime:2.3.3 etc.) but Auto-WEKA do not have the binding information of the implementation classes at runtime. A simple solution is to add the binding information (in jaxb.properties) with the class XmlSerializable.